# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Resume Analyzer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ResumeUploadInput {
  userId: string;
  fileName: string;
}

export type ResumeStatus = typeof ResumeStatus[keyof typeof ResumeStatus];


export const ResumeStatus = {
  pending: 'pending',
  analyzing: 'analyzing',
  done: 'done',
  failed: 'failed',
} as const;

export interface Resume {
  id: number;
  userId: string;
  fileName: string;
  uploadedAt: string;
  status: ResumeStatus;
  /** @nullable */
  atsScore?: number | null;
  /** @nullable */
  extractedText?: string | null;
  /** @nullable */
  skills?: string | null;
  /** @nullable */
  missingSkills?: string | null;
  /** @nullable */
  suggestions?: string | null;
  /** @nullable */
  education?: string | null;
  /** @nullable */
  experience?: string | null;
  /** @nullable */
  projects?: string | null;
}

export interface DashboardStats {
  totalResumes: number;
  /** @nullable */
  averageAtsScore: number | null;
  /** @nullable */
  bestScore: number | null;
  analyzedCount: number;
  recentActivity?: Resume[];
}

export interface SkillGapItem {
  skill: string;
  count: number;
}

export type ListResumesParams = {
userId: string;
};

export type GetDashboardStatsParams = {
userId: string;
};

export type GetSkillGapsParams = {
userId: string;
};