# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Resume Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { Resume } from './resume';

export interface DashboardStats {
  totalResumes: number;
  /** @nullable */
  averageAtsScore: number | null;
  /** @nullable */
  bestScore: number | null;
  analyzedCount: number;
  recentActivity?: Resume[];
}