# lib/api-zod/src/generated/types/resume.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Resume Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { ResumeStatus } from './resumeStatus';

export interface Resume {
  id: number;
  userId: string;
  fileName: string;
  uploadedAt: string;
  status: ResumeStatus;
  /** @nullable */
  atsScore?: number | null;
  /** @nullable */
  extractedText?: string | null;
  /** @nullable */
  skills?: string | null;
  /** @nullable */
  missingSkills?: string | null;
  /** @nullable */
  suggestions?: string | null;
  /** @nullable */
  education?: string | null;
  /** @nullable */
  experience?: string | null;
  /** @nullable */
  projects?: string | null;
}