# lib/api-zod/src/generated/types/resumeUploadInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Resume Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface ResumeUploadInput {
  userId: string;
  fileName: string;
}